backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
djoser==2.2.0
djangorestframework==3.16.1 # Upgrade for Django 6
drf-spectacular==0.27.2
pytest-django==4.8.0
pytest==8.2.0
pytest-cov==5.0.0
sqlparse==0.5.5
psycopg2-binary==2.9.11
